move more code to presentation
